feat: add Stripe checkout with discounts & shipping
- Introduce Stripe checkout session creation supporting discounts and shipping
- Refactor cart page to use new CartPage component
- Add sync-discount route for Stripe coupons & promo codes
- Add fields: firstTimeCustomer, updatedBySync (hidden), stripePromoId (hidden), redemptions (read-only)
- Update Sanity config for custom publish/delete actions to reset updatedBySync
- Add supporting files for Stripe and Sanity actions
