+++ src/app/(client)/success/page.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SuccessPage = () => {
+  return (
+    <div>SuccessPage</div>
+  )
+}
+
+export default SuccessPage
